image safety mark one

+++ src/app/api/deleteMedia/[userId]/[fileName]/route.ts
@@ -1,6 +1,7 @@
 import { NextRequest, NextResponse } from "next/server";
 import fs from "fs";
 import path from "path";
+import prismadb from "@/lib/prismadb";
 
 export async function DELETE(
   req: NextRequest,
@@ -14,6 +15,20 @@ export async function DELETE(
   }
 
   try {
+    // Remove the representation of the image from the database first
+    const image = await prismadb.image.findFirst({
+      where: { fileName },
+    });
+
+    if (!image) {
+      return NextResponse.json({ error: "Image not found" });
+    }
+
+    await prismadb.image.delete({
+      where: { id: image.id },
+    });
+
+    // Then, remove the image file from the server
     const filePath = path.join("./public/uploads", userId, fileName);
     fs.unlinkSync(filePath);
